fix more cyclic imports
