+++ src/format/asciidoc/format-asciidoc.ts
@@ -1,9 +1,8 @@
 /*
-* format-asciidoc.ts
-*
-* Copyright (C) 2020-2022 Posit Software, PBC
-*
-*/
+ * format-asciidoc.ts
+ *
+ * Copyright (C) 2020-2022 Posit Software, PBC
+ */
 
 import { Format } from "../../config/types.ts";
 
@@ -11,15 +10,17 @@ import { mergeConfigs } from "../../core/config.ts";
 import { resolveInputTarget } from "../../project/project-index.ts";
 import {
   BookChapterEntry,
-  bookOutputStem,
   BookPart,
-  kBookAppendix,
-  kBookChapters,
 } from "../../project/types/book/book-config.ts";
 import {
   bookConfig,
+  bookOutputStem,
   isBookIndexPage,
 } from "../../project/types/book/book-shared.ts";
+import {
+  kBookAppendix,
+  kBookChapters,
+} from "../../project/types/book/book-constants.ts";
 import { join, relative } from "path/mod.ts";
 
 import { plaintextFormat } from "../formats-shared.ts";
