+++ src/project/types/book/book-constants.ts
@@ -1,0 +1,19 @@
+/*
+ * book-constants.ts
+ *
+ * Copyright (C) 2020-2023 Posit Software, PBC
+ */
+
+export const kBookChapters = "chapters";
+export const kBookAppendix = "appendices";
+export const kBookReferences = "references";
+export const kBookRender = "render";
+export const kBookOutputFile = "output-file";
+export const kBookRepoActions = "repo-actions";
+export const kBookSharing = "sharing";
+export const kBookDownloads = "downloads";
+export const kBookTools = "tools";
+export const kBookSearch = "search";
+export const kBookItemChapter = "chapter";
+export const kBookItemAppendix = "appendix";
+export const kBookItemPart = "part";
